Allows users to select a default query
Adds the plugin `DefaultUserQueryPlugin` which allows users to select a
personal default query.

When the user visits the ticket page without specififying a query a
check is made to see whether the user has a saved default query.  If
that is the case that query is executed instead of the one specified in
the trac config as `query.default_query`.

In order for this to be containable in a plugin without modifying any
other #define code or templates this plugin replaces all links to
`/[projectroot]/query' that don't have a query string with a links to
the users default query.

Arguably this kind of `DOM` modification should be done with a
`genshi.filters.transform.Transformer`, but I couldn't get that to
update the link in the ribbon so I decided on just using jQuery instead.

Because of the link in the ribbon it is expected that there is at least
one link to `/[projectroot]/query` on any given #define page, as a result
the `filter_stream` method of this plugin will be executed on each page
load.

A link is added to the ribbon which saves the current query as the
default.  The link triggers an ajax call that sends the query string of
the current query page back to the plugin where it is saved.  Currently
the only feedback is a slight blinking effect that may be too weak.
Ideally the `add_notice` method should be used, but that requires a page
reload.

This plugin does not require any schema changes.  The table
`session_attribute` is used for storing the users default queries under
the key `default_user_query`.

Fixes #3703

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@20515 484df14f-75d2-4b3f-8ba0-490e05069a01