Redesign and refactoring of defaultuserqueryplugin
There was no need to execute sql on the `session_attribute` table, and
it turned out writing to the database followed by a redirect can easily
lead to hanging transactions, so the sql statements have been removed
in favour of just manipulating the `req.session` dict and letting the
framework persist it.  I simply did not know that `req.session` was
automatically persisted.

The user interface has been improved.  In [20515] a link was placed in
the ribbon for changing the default query.  This had the downside that
it wasn't part of the query form so the current querystring had to be
used as the default query.  Changing the default query was a two-step
action.  First building the query and pressing `update`, then pressing
`Set default query`.

With this commit the link is replaced with a button in the query form.
The button posts the content of the form to this plugin and builds a
query using the form data.  The private method
`trac.ticket.query.QueryModule._get_constraints` had to be called for
this to work.  Once the query has been saved as default the request is
redirected to the query page and handled as if the user had pressed
`update`.

The `HTTP` call made to this plugin is no longer an ajax call, just a
normal form post.  The `action` of the form has to be changed prior to
submitting to prevent the post from being picked up by the query module.
That redirection is made with an `onclick` attribute which is a little
bit hacky, but on injected html I think it's acceptable.

After the modifications above hade been made the javascript file
introduced in [20515] had shrunk to a one-liner, it was removed and the
javascript needed is injected inlined into the html template instead.

Fixes #3703

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@20956 484df14f-75d2-4b3f-8ba0-490e05069a01