Replaces the onclick attribute with a jQuery selector.
An `onclick` attribute was added to the button `Set as default` in
[20956].  Even though I apologized for it in the commit message the
changeset did not pass quality assurance.  The javascript has been moved
out of the `onclick` attribute.

Also adds some spaces in and around the injected button for consistency
with the rest of the user interface.

The script is still technically inline, but insite a `script` tag in the
header rather than in an attribute of the button.  I'm still not sure
whether it should have been in a separate javascript file, on one hand
it would cost an extra HTTP request per page load, but on the other hand
it wouldn't be served from our servers so it wouldn't really be our
problem.  I remember #1364 which called for inlining a short javascript
file so I did the same with this plugin.

Fixes #3703

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@21121 484df14f-75d2-4b3f-8ba0-490e05069a01