Replaces the form button for changing default query with a ribbon link
This was the plan from the start but the ribbon link was changed to a
form button in [20956] due to a technical reason.  Once it was in the
form I found I liked it better that way and kept it there.

When it showed up on the staging environment several team members
disapproved of the user experience and it was decided to move it to the
ribbon.

Moving the button/link out of the form didn't require more changes to
the event handler than simply calling $.submit manually at the end.

I'm not sure how the items in the ribbon are ordered, but the link ended
up exactly where it should be (next to "save these filters") right away,
so I won't try to figure that out unless it starts moving abuout during
review and testing.

Fixes #3903. Took 1 hour.

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@21788 484df14f-75d2-4b3f-8ba0-490e05069a01